requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
requests>=2.31.0
textstat>=0.7.7
nltk>=3.8.1
python-dateutil>=2.8.2
